Fix issue with min/max temperatures in precis forcast
As reported by @Keith-Pembleton, there were issues with the min
temperature > max temperature in precis forecasts.

This was due to improper handling of the conversion from factor to
numeric in `get_precis_forecast()`. The same issue affected
precipitation values (upper, lower and probability).

This commit also fixes another bug where the format of the date/times
in the precis forecast were not returned as POSIXct.

This commit also fixes a bug where the upper precipitation limit was
not returned as numeric.